[MOD/#35] 채팅 패키지 및 네비게이션 수정
