+++ app/src/main/java/org/sopt/linkareer/feature/chattingroom/component/OtherUserChat.kt
@@ -1,4 +1,4 @@
-package org.sopt.linkareer.feature.chatting.component
+package org.sopt.linkareer.feature.chattingroom.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -6,11 +6,10 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -22,38 +21,34 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.sopt.linkareer.R
 import org.sopt.linkareer.core.designsystem.theme.Blue50
-import org.sopt.linkareer.core.designsystem.theme.Gray300
 import org.sopt.linkareer.core.designsystem.theme.Gray600
 import org.sopt.linkareer.core.designsystem.theme.Gray900
+import org.sopt.linkareer.core.designsystem.theme.LINKareerAndroidTheme
 import org.sopt.linkareer.core.designsystem.theme.defaultLINKareerTypography
 
 @Composable
-fun MyReplyChat(
-    sender: String,
-    receivedMessage: String,
-    replyMessage: String,
+fun OtherUserChat(
+    nickName: String,
+    isChecked: Boolean,
+    jobCategory: String,
+    imageUrl: String,
+    sendMessage: String,
     timestamp: String,
 ) {
-    Column(
-        modifier =
-            Modifier
-                .padding(end = 8.dp),
-        horizontalAlignment = Alignment.End,
-    ) {
+    Column {
+        OtherUserChatProfile(
+            nickName = nickName,
+            isChecked = isChecked,
+            jobCategory = jobCategory,
+            imageUrl = imageUrl,
+        )
         Row(
             modifier =
                 Modifier
-                    .wrapContentWidth(),
+                    .fillMaxWidth()
+                    .padding(start = 30.dp),
             verticalAlignment = Alignment.Bottom,
         ) {
-            Text(
-                text = timestamp,
-                style = defaultLINKareerTypography.body13R11,
-                color = Gray600,
-                modifier =
-                    Modifier
-                        .padding(end = 4.dp),
-            )
             Column(
                 modifier =
                     Modifier
@@ -64,27 +59,19 @@ fun MyReplyChat(
                 verticalArrangement = Arrangement.spacedBy(space = 8.dp, alignment = Alignment.Bottom),
             ) {
                 Text(
-                    text = sender + " " + stringResource(R.string.chatroom_apply_to_sender),
-                    style = defaultLINKareerTypography.body5B11,
-                    color = Gray900,
-                )
-                Text(
-                    text = receivedMessage,
-                    style = defaultLINKareerTypography.label3M11,
-                    color = Gray600,
-                )
-
-                HorizontalDivider(
-                    color = Gray300,
-                    thickness = 1.dp,
-                )
-
-                Text(
-                    text = replyMessage,
+                    text = sendMessage,
                     style = defaultLINKareerTypography.body8M13,
                     color = Gray900,
                 )
             }
+            Text(
+                text = timestamp,
+                style = defaultLINKareerTypography.body13R11,
+                color = Gray600,
+                modifier =
+                    Modifier
+                        .padding(start = 4.dp),
+            )
         }
         Image(
             painter = painterResource(R.drawable.ic_chatting_like_inactive_25),
@@ -101,11 +88,15 @@ fun MyReplyChat(
     backgroundColor = 0xFFFFFFFF,
 )
 @Composable
-fun MyReplyChatPreview() {
-    MyReplyChat(
-        sender = "nn",
-        receivedMessage = "origin message",
-        replyMessage = "text message",
-        timestamp = "18:33",
-    )
+fun OtherUserChatPreview() {
+    LINKareerAndroidTheme {
+        OtherUserChat(
+            nickName = "무관심한 맥",
+            isChecked = true,
+            jobCategory = "현대 자동차",
+            imageUrl = "",
+            sendMessage = "굳이 꾸밀 필요없습니다.",
+            timestamp = "18:33",
+        )
+    }
 }
