+++ app/src/main/java/org/sopt/linkareer/feature/main/MainScreen.kt
@@ -21,7 +21,8 @@ import org.sopt.linkareer.core.designsystem.theme.Gray900
 import org.sopt.linkareer.core.designsystem.theme.LINKareerTheme
 import org.sopt.linkareer.core.designsystem.theme.White
 import org.sopt.linkareer.core.util.NoRippleInteraction
-import org.sopt.linkareer.feature.chatting.navigation.chattingGraph
+import org.sopt.linkareer.feature.chattinghome.navigation.chattingHomeGraph
+import org.sopt.linkareer.feature.chattingroom.navigation.chattingRoomGraph
 import org.sopt.linkareer.feature.home.navigation.homeNavGraph
 
 @Composable
@@ -64,8 +65,11 @@ fun MainScreen(
                     paddingValues = paddingValues,
                     navHostController = navigator.navController,
                 )
-                chattingGraph(
+                chattingHomeGraph(
                     paddingValues = paddingValues,
+                    navHostController = navigator.navController,
+                )
+                chattingRoomGraph(
                     navHostController = navigator.navController,
                 )
             }
